% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aat.R
\name{.aat_parse_itl}
\alias{.aat_parse_itl}
\title{Parse .itl.csv file (raw responses)}
\usage{
.aat_parse_itl(file_path, rel_path, code, date, ambiguous_items, control_items)
}
\description{
Parse .itl.csv file (raw responses)
}
\keyword{internal}
