% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aat.R
\name{.aat_parse_rsl}
\alias{.aat_parse_rsl}
\title{Parse .rsl.csv file (computed results)}
\usage{
.aat_parse_rsl(file_path, rel_path, code, date)
}
\description{
Parse .rsl.csv file (computed results)
}
\keyword{internal}
