% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppt.R
\name{.create_single_profile_plot}
\alias{.create_single_profile_plot}
\title{Create single PPPT profile plot}
\usage{
.create_single_profile_plot(data, title, color, show_legend, label)
}
\arguments{
\item{data}{Long format data for one line}

\item{title}{Plot title}

\item{color}{Line color}

\item{show_legend}{Show legend}

\item{label}{Line label}
}
\value{
Plotly object
}
\description{
Create single PPPT profile plot
}
\keyword{internal}
