% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppt.R
\name{.empty_pppt_tibble}
\alias{.empty_pppt_tibble}
\title{Create empty PPPT tibble with correct structure}
\usage{
.empty_pppt_tibble(with_group = FALSE)
}
\arguments{
\item{with_group}{Whether to include group column}
}
\value{
Empty tibble
}
\description{
Create empty PPPT tibble with correct structure
}
\keyword{internal}
