% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppt.R
\name{.extract_group_from_path}
\alias{.extract_group_from_path}
\title{Extract group from file path}
\usage{
.extract_group_from_path(file_path, root, group_names)
}
\arguments{
\item{file_path}{Full file path}

\item{root}{Root directory}

\item{group_names}{Group names to look for}
}
\value{
Group name or "no_group"
}
\description{
Extract group from file path
}
\keyword{internal}
