% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppt.R
\name{.pppt_to_long}
\alias{.pppt_to_long}
\title{Convert PPPT data to long format for plotting}
\usage{
.pppt_to_long(data)
}
\arguments{
\item{data}{Wide format PPPT data}
}
\value{
Long format tibble
}
\description{
Convert PPPT data to long format for plotting
}
\keyword{internal}
