% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klawa.R
\name{extract_and_check_code}
\alias{extract_and_check_code}
\title{Extract a unique participant code and detect conflicts}
\usage{
extract_and_check_code(x, pattern = "(\\\\d{4}[A-Za-zÄÖÜäöüß]{4})")
}
\arguments{
\item{x}{String (path/filename).}

\item{pattern}{Regex for the code (default: 4 digits + 4 letters, including German umlauts).}
}
\value{
Single code (character), \code{NA}, or \code{"CODE_CONFLICT"}.
}
\description{
Extract a unique participant code and detect conflicts
}
