% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the musicAnalysis Shiny app}
\usage{
launch_app()
}
\value{
Invisibly TRUE if app directory is found; otherwise an error is thrown.
}
\description{
Opens an interactive UI to run common workflows:
\itemize{
\item Scan KLAWA PDFs
\item Parse Musical Experience (LimeSurvey CSV)
\item Merge datasets by \code{code}
\item Adjust basic options
}
}
