% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppt.R
\name{.create_overlaid_profile_plot}
\alias{.create_overlaid_profile_plot}
\title{Create overlaid PPPT profile plot}
\usage{
.create_overlaid_profile_plot(data, color_by, line_color, show_legend, title)
}
\arguments{
\item{data}{Long format data for multiple lines}

\item{color_by}{How to color lines}

\item{line_color}{Default line color}

\item{show_legend}{Show legend}

\item{title}{Plot title}
}
\value{
Plotly object
}
\description{
Create overlaid PPPT profile plot
}
\keyword{internal}
