% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppt.R
\name{pppt_analyze_structure}
\alias{pppt_analyze_structure}
\title{Analyze PPPT folder structure}
\usage{
pppt_analyze_structure(root)
}
\arguments{
\item{root}{Path to root folder}
}
\value{
List with structure information
}
\description{
Scans folder to identify PPPT files and their organization
}
